requirements.txt
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-embed-video==1.4.0
django-heroku==0.0.0
djangorestframework==3.12.2
gunicorn==20.0.4
html5lib==1.1
httplib2==0.18.1
idna==2.10
layoutswitcherlib==0.8.16
Pillow==8.1.0
requests==2.25.1
requests-file==1.5.1
sqlparse==0.4.1
urllib3==1.26.3
whitenoise==5.2.0
